Add RingDecomposerLib, integrate
Add RingDecomposerLib
- From https://github.com/rareylab/RingDecomposerLib, commit eb4cbdf
- Papers to cite:
  - Kolodzik, A.; Urbaczek, S.; Rarey, M. Unique Ring Families: A Chemically
    Meaningful Description of Molecular Ring Topologies. J. Chem. Inf. Model.,
    2012, 52 (8), pp 2013–2021
  - Flachsenberg, F.; Andresen, N.; Rarey, M. RingDecomposerLib: An Open-Source
    implementation of Unique Ring Families and Other Cycle Bases. J. Chem. Inf.
    Model., 2017, 57 (2), pp 122–126

- License: BSD 3-clasue (New)
  - Must include copy of license and copyright notice in all distributions of
    the software

Integrate RDL
- Appropriately modify CMakeLists
- Add wrappers for ANSI C pointer and memory handling (to get C++ convenience)

Various
- clang-tidy improvements

+++ source/DistanceGeometry/generateConformation.cpp
@@ -40,7 +40,7 @@ Delib::PositionCollection convertToPositionCollection(
 Delib::PositionCollection convertToPositionCollection(
   const dlib::matrix<double, 0, 1>& vectorizedPositions
 ) {
-  const unsigned dimensionality = 4;
+  const long dimensionality = 4;
   assert(vectorizedPositions.size() % dimensionality == 0);
 
   Delib::PositionCollection positions;
@@ -104,8 +104,8 @@ ChiralityConstraint propagate(
   lowerMatrix.diagonal().setZero();
   upperMatrix.diagonal().setZero();
 
-  for(unsigned i = 0; i < 4; i++) {
-    for(unsigned j = i + 1; j < 4; j++) {
+  for(long i = 0; i < 4; i++) {
+    for(long j = i + 1; j < 4; j++) {
       const double& lowerBound = bounds.lowerBound(
         prototype.indices.at(i),
         prototype.indices.at(j)
@@ -265,7 +265,7 @@ std::list<Delib::PositionCollection> runDistanceGeometry(
         ) < 0.5
       ) {
         // Invert y coordinates
-        for(unsigned i = 0; i < DGData.distanceBounds.N; i++) {
+        for(long i = 0; i < DGData.distanceBounds.N; i++) {
           dlibPositions(4 * i + 1) *= -1;
         }
       }
